fix: harden pinned tooltip and factor the chart per review
Pin now tracks the period key so a refetch cannot strand a stale card
(auto-unpins if the period disappears); pin state is read via a ref so
toggling it no longer rebuilds the recharts tree; the pin/hover state
machine moves to usePinnedChartTooltip and TooltipItem to its own file;
display-info types live once in @autumn/shared; the plan group-by
branch uses groupValueLabel; tooltip link resolution is a tested pure
function.
